feat: caf pool contracts

+++ contracts/tokenomics/CAFToken.sol
@@ -9,16 +9,16 @@ import {Ownable} from "@openzeppelin/contracts/access/Ownable.sol";
 contract CAFToken is ICAFToken, ERC20, Ownable {
     constructor(
         address initialOwner,
-        uint256 initialSupply
+        uint128 initialSupply
     ) Ownable(initialOwner) ERC20("CaFi", "CAF") {
         _mint(msg.sender, initialSupply);
     }
 
-    function mint(address to, uint256 amount) public onlyOwner {
+    function mint(address to, uint128 amount) public onlyOwner {
         _mint(to, amount);
     }
 
-    function burn(uint256 amount) public {
+    function burn(uint128 amount) public {
         _burn(msg.sender, amount);
     }
 }
